Rework presentations to be scalable
