+++ frontend/src/components/pages/CreatorProfile/PresentationForm.tsx
@@ -23,7 +23,6 @@ import { Book } from "@material-ui/icons";
 import React, { useContext, useEffect, useState } from "react";
 
 import CreatorProfileContext from "../../../contexts/CreatorProfileContext";
-import { Option } from "../../../types/BookTypes";
 import AddMultiSelect from "../CreateReview/AddMultiSelect";
 import CreatorInputField from "./CreatorInputField";
 import WorkshopFormContainer from "./WorkshopFormContainer";
@@ -32,74 +31,18 @@ interface PresentationFormProps {
   submitted: boolean;
 }
 
-interface MultiSelectHandler {
-  newValue: Option[];
-  field: "offeredLocations" | "preferredGradeLevel";
-}
-
-interface CheckboxGroupHandler {
-  newValue: StringOrNumber[];
-  field: "languages";
-}
-
-interface InputHandler {
-  newValue: string;
-  field: "inPersonDeliveryFee" | "virtualDeliveryFee" | "otherReadingLanguages";
-}
-
-interface RadioGroupHandler {
-  newValue: string;
-  field: "booksPurchasedAndAutoGraphed";
-}
-
 const PresentationForm = ({
   submitted,
 }: PresentationFormProps): React.ReactElement => {
   const { creatorProfile, setCreatorProfile } = useContext(
     CreatorProfileContext,
   );
-  /**
-   * @param newOptionsSelected handle updating context
-   * @param field string of key in context value
-   */
-  const handleFormInputChange = (
-    keyValue:
-      | MultiSelectHandler
-      | InputHandler
-      | CheckboxGroupHandler
-      | RadioGroupHandler,
-  ) => {
-    const { newValue, field } = keyValue;
-    const creatorProfileObj = { ...creatorProfile };
-    creatorProfileObj[field] = newValue;
-    setCreatorProfile(creatorProfileObj);
-  };
-
-  const [offeredLocationsOptions, setOfferedLocationsOptions] = useState<
-    Option[]
-  >([]);
-  const [preferredGradeLevelOptions, setPreferredGradeLevelOptions] = useState<
-    Option[]
-  >([]);
 
   const [isReadingSectionOpen, setIsReadingSectionOpen] = useState(false);
   const [isWorkshopSectionOpen, setIsWorkshopSectionOpen] = useState(false);
 
   // Set inital multiselection options
-  useEffect(() => {
-    // TODO: Replace with DB calls
-    setOfferedLocationsOptions([
-      { label: "Schools", value: "Schools" },
-      { label: "Libraries", value: "Libraries" },
-    ]);
 
-    // TODO: Replace with DB calls
-    setPreferredGradeLevelOptions([
-      { label: "Primary School", value: "Primary School" },
-      { label: "Middle Grade", value: "Primary Grade" },
-      { label: "High School", value: "High School" },
-    ]);
-  }, []);
 
   return (
     <Flex flex="1" direction="column" justify="start">
@@ -115,6 +58,7 @@ const PresentationForm = ({
         title="Readings"
         isOpen={isReadingSectionOpen}
         setIsOpen={setIsReadingSectionOpen}
+        index={0}
         icon={
           <Icon w="24px" h="24px" viewBox="0 0 18 18" fill="none" radius="50%">
             <path
@@ -133,156 +77,15 @@ const PresentationForm = ({
             />
           </Icon>
         }
-      >
-        <>
-          <AddMultiSelect
-            id="offeredLocations"
-            label="Offered locations"
-            placeholder="Select or add your own option"
-            optionsSelected={creatorProfile?.offeredLocations || []}
-            setOptionsSelected={(newOptionsSelected: Option[]) => {
-              handleFormInputChange({
-                newValue: newOptionsSelected,
-                field: "offeredLocations",
-              });
-            }}
-            options={offeredLocationsOptions}
-            setOptions={setOfferedLocationsOptions}
-            allowAddOption
-            allowMultiSelectOption
-            maxWidth="80%"
-          />
-          <AddMultiSelect
-            id="preferredGradeLevel"
-            label="Preferred grade level"
-            placeholder="Select or add your own option"
-            optionsSelected={creatorProfile?.preferredGradeLevel || []}
-            setOptionsSelected={(newOptionsSelected: Option[]) => {
-              handleFormInputChange({
-                newValue: newOptionsSelected,
-                field: "preferredGradeLevel",
-              });
-            }}
-            options={preferredGradeLevelOptions}
-            setOptions={setPreferredGradeLevelOptions}
-            allowAddOption
-            allowMultiSelectOption
-            maxWidth="80%"
-          />
-          <CreatorInputField
-            name="Preferred audience size"
-            value={creatorProfile?.preferredAudienceSize}
-            field="preferredAudienceSize"
-            error={submitted}
-            selectOptions={["0-5", "6-10", "11-15", "16-20", "21-25"]}
-            width="33%"
-          />
-          <FormControl isRequired isInvalid={submitted}>
-            <FormLabel mb="1" mt="3">
-              How do you deliver readings?
-            </FormLabel>
-            <Flex>
-              <Button> In-person </Button>
-              <Input
-                name="In-person delivery fee"
-                placeholder="Enter fee for in-person"
-                value={creatorProfile?.inPersonDeliveryFee}
-                onChange={(e) => {
-                  handleFormInputChange({
-                    newValue: e.target.value,
-                    field: "inPersonDeliveryFee",
-                  });
-                }}
-              />
-            </Flex>
-            <Flex>
-              <Button> Virtual </Button>
-              <Input
-                name="Virtual delivery fee"
-                placeholder="Enter fee for virtual"
-                value={creatorProfile?.virtualDeliveryFee}
-                onChange={(e) => {
-                  handleFormInputChange({
-                    newValue: e.target.value,
-                    field: "virtualDeliveryFee",
-                  });
-                }}
-              />
-            </Flex>
-          </FormControl>
-          <CreatorInputField
-            name="Please describe any AV materials or special equipment required."
-            value={creatorProfile?.equipmentRequired}
-            field="equipmentRequired"
-            error={submitted}
-            width="33%"
-          />
-          <FormControl>
-            <FormLabel mb="1" mt="3">
-              Please list the languages your readings are avaliable in:
-            </FormLabel>
-            <Flex direction="column">
-              <CheckboxGroup
-                value={creatorProfile?.languages || []}
-                onChange={(newValueSelected) => {
-                  handleFormInputChange({
-                    newValue: newValueSelected,
-                    field: "languages",
-                  });
-                }}
-              >
-                <Checkbox value="english"> English </Checkbox>
-                <Checkbox value="french"> French </Checkbox>
-                <Checkbox value="other"> Other </Checkbox>
-              </CheckboxGroup>
-            </Flex>
-            <Flex alignContent="center">
-              <FormLabel w="270px">If Other, please specify:</FormLabel>
-              <Input
-                placeholder="Specify other languages"
-                onChange={(e) => {
-                  handleFormInputChange({
-                    newValue: e.target.value,
-                    field: "otherReadingLanguages",
-                  });
-                }}
-              />
-            </Flex>
-            <FormLabel />
-          </FormControl>
-          <FormControl>
-            <FormLabel>
-              {" "}
-              For in-person visits, do you bring copies of your books for
-              students to purchase and have autographed?
-            </FormLabel>
-            <RadioGroup
-              onChange={(newValueSelected) => {
-                handleFormInputChange({
-                  newValue: newValueSelected,
-                  field: "booksPurchasedAndAutoGraphed",
-                });
-              }}
-              value={creatorProfile?.booksPurchasedAndAutoGraphed}
-            >
-              <Flex direction="column">
-                <Radio value="yes"> Yes </Radio>
-                <Radio value="no"> No </Radio>
-              </Flex>
-            </RadioGroup>
-          </FormControl>
-          <FormControl isRequired>
-            <FormLabel> Describe the contents of your readings(s)</FormLabel>
-            <Textarea />
-          </FormControl>
-        </>
-      </WorkshopFormContainer>
-      <Divider size="4px" />
+        submitted
+      />
+      <Divider size="4px" m="24px" />
       {/* Workshops section */}
       <WorkshopFormContainer
         title="Workshops"
         isOpen={isWorkshopSectionOpen}
         setIsOpen={setIsWorkshopSectionOpen}
+        index={1}
         icon={
           <Icon w="24px" h="24px" viewBox="0 0 18 18" fill="none" radius="50%">
             <path
@@ -301,150 +104,8 @@ const PresentationForm = ({
             />
           </Icon>
         }
-      >
-        <>
-          <AddMultiSelect
-            id="offeredLocations"
-            label="Offered locations"
-            placeholder="Select or add your own option"
-            optionsSelected={creatorProfile?.offeredLocations || []}
-            setOptionsSelected={(newOptionsSelected: Option[]) => {
-              handleFormInputChange({
-                newValue: newOptionsSelected,
-                field: "offeredLocations",
-              });
-            }}
-            options={offeredLocationsOptions}
-            setOptions={setOfferedLocationsOptions}
-            allowAddOption
-            allowMultiSelectOption
-            maxWidth="80%"
-          />
-          <AddMultiSelect
-            id="preferredGradeLevel"
-            label="Preferred grade level"
-            placeholder="Select or add your own option"
-            optionsSelected={creatorProfile?.preferredGradeLevel || []}
-            setOptionsSelected={(newOptionsSelected: Option[]) => {
-              handleFormInputChange({
-                newValue: newOptionsSelected,
-                field: "preferredGradeLevel",
-              });
-            }}
-            options={preferredGradeLevelOptions}
-            setOptions={setPreferredGradeLevelOptions}
-            allowAddOption
-            allowMultiSelectOption
-            maxWidth="80%"
-          />
-          <CreatorInputField
-            name="Preferred audience size"
-            value={creatorProfile?.preferredAudienceSize}
-            field="preferredAudienceSize"
-            error={submitted}
-            selectOptions={["0-5", "6-10", "11-15", "16-20", "21-25"]}
-            width="33%"
-          />
-          <FormControl isRequired isInvalid={submitted}>
-            <FormLabel mb="1" mt="3">
-              How do you deliver readings?
-            </FormLabel>
-            <Flex>
-              <Button> In-person </Button>
-              <Input
-                name="In-person delivery fee"
-                placeholder="Enter fee for in-person"
-                value={creatorProfile?.inPersonDeliveryFee}
-                onChange={(e) => {
-                  handleFormInputChange({
-                    newValue: e.target.value,
-                    field: "inPersonDeliveryFee",
-                  });
-                }}
-              />
-            </Flex>
-            <Flex>
-              <Button> Virtual </Button>
-              <Input
-                name="Virtual delivery fee"
-                placeholder="Enter fee for virtual"
-                value={creatorProfile?.virtualDeliveryFee}
-                onChange={(e) => {
-                  handleFormInputChange({
-                    newValue: e.target.value,
-                    field: "virtualDeliveryFee",
-                  });
-                }}
-              />
-            </Flex>
-          </FormControl>
-          <CreatorInputField
-            name="Please describe any AV materials or special equipment required."
-            value={creatorProfile?.equipmentRequired}
-            field="equipmentRequired"
-            error={submitted}
-            width="33%"
-          />
-          <FormControl>
-            <FormLabel mb="1" mt="3">
-              Please list the languages your readings are avaliable in:
-            </FormLabel>
-            <Flex direction="column">
-              <CheckboxGroup
-                value={creatorProfile?.languages || []}
-                onChange={(newValueSelected) => {
-                  handleFormInputChange({
-                    newValue: newValueSelected,
-                    field: "languages",
-                  });
-                }}
-              >
-                <Checkbox value="english"> English </Checkbox>
-                <Checkbox value="french"> French </Checkbox>
-                <Checkbox value="other"> Other </Checkbox>
-              </CheckboxGroup>
-            </Flex>
-            <Flex alignContent="center">
-              <FormLabel w="270px">If Other, please specify:</FormLabel>
-              <Input
-                placeholder="Specify other languages"
-                onChange={(e) => {
-                  handleFormInputChange({
-                    newValue: e.target.value,
-                    field: "otherReadingLanguages",
-                  });
-                }}
-              />
-            </Flex>
-            <FormLabel />
-          </FormControl>
-          <FormControl>
-            <FormLabel>
-              {" "}
-              For in-person visits, do you bring copies of your books for
-              students to purchase and have autographed?
-            </FormLabel>
-            <RadioGroup
-              onChange={(newValueSelected) => {
-                handleFormInputChange({
-                  newValue: newValueSelected,
-                  field: "booksPurchasedAndAutoGraphed",
-                });
-              }}
-              value={creatorProfile?.booksPurchasedAndAutoGraphed}
-            >
-              <Flex direction="column">
-                <Radio value="yes"> Yes </Radio>
-                <Radio value="no"> No </Radio>
-              </Flex>
-            </RadioGroup>
-          </FormControl>
-          <FormControl isRequired>
-            <FormLabel> Describe the contents of your readings(s)</FormLabel>
-            <Textarea />
-          </FormControl>
-        </>
-      </WorkshopFormContainer>
+        submitted
+      />
       <Divider />
       <Button color="blue.400" w="30%" justifyContent="start" p="0">
         <Flex justifyContent="space-between">
