+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@chakra-ui/utils": "^2.0.14",
     "@fontsource/dm-sans": "^4.5.0"
   }
 }
